requirements.txt
opcua==0.98.1
cryptography==2.2.2
pyserial==3.4
PyYAML==3.13
requests==2.19.1
Flask==1.0.2
Flask-Cors==3.0.6